Move Adaptive S-Foils to T70XWing file, cleanup
