+++ Assets/Scripts/Model/Content/SecondEdition/Upgrades/Talent/ForTheCause.cs
@@ -1,14 +1,11 @@
-using Content;
 using System;
 using System.Collections.Generic;
-using System.Linq;
-using Upgrade;
-using UpgradesList.SecondEdition;
+using ActionsList;
+using Content;
 using Ship;
 using SubPhases;
 using Tokens;
-using ActionsList;
-using Abilities;
+using Upgrade;
 
 namespace UpgradesList.SecondEdition
 {
